Crate ggpf

General Game Playing Framework.

This crate introduces common APIs for games and policies to explore general game playing algorithms.

Among classical tree search-based methods there are attempts to re-implement AlphaZero and MuZero. AlphaZero has been tested on a toy example, but MuZero is not yet successful.

Binaries

Several binaries can be used to test the project:

evaluate: test two policies against each other.

generate: generate self-play games for AlphaZero and MuZero.

generate: generate self-play games for AlphaZero and MuZero.

ui: visualize PUCT-based policies (AlphaZero/MuZero) in a duel against PPA.

perf: test raw games generation performance.

Modules

deep
Features for neural network-based policies.

game
General game traits and implementations.

policies
General policy traits and implementations.

settings
General game and playout settings.