Module ggpf::deep::evaluator

Neural networks evaluators.

Functions

dynamics
Dynamics evaluator

dynamics_evaluator_single
Dynamics evaluator - single batch

dynamics_task
Dynamics task

prediction
Prediction evaluator

prediction_evaluator_single
Evaluates a game state for PUCT - single batch

prediction_task
Prediction task

representation
Representation evaluator

representation_evaluator_single
State to representation for Muz - single batch

representation_task
Representation task

Type Definitions

DynamicsEvaluatorChannel
Takes a tensor and a way to send back the inference result for the dynamics network.

PredictionEvaluatorChannel
Takes a tensor and a way to send back the inference result for the prediction network.

RepresentationEvaluatorChannel
Takes a tensor and a way to send back the inference result for the representation network.