Module ggpf::deep::self_play

Self-play generation.

Asynchronous evaluation functions.

For each kind of evaluation there is an evaluator that can be provided to PUCT/Muz and an associated evaluator task that sends the request to the GPU while batching them.

Structs

GameHistoryEntry
Game history data generated from self-play

Functions

alphazero_game_generator
AlphaZero self-play games generator

muzero_game_generator
MuZero self-play games generator Spawn several tasks (number according to settings) that performs self-play games using the MuZero policy, sending them in the output_chan channel.