Module ggpf::game

General game traits and implementations.

Modules

breakthrough
Breakthrough game implementation

hashcode_20
Google Hashcode 2020 problem.

meta
Games that takes other games as an input.

misere_breakthrough
Misère breakthrough

openai
Open AI Gym interface.

weak_schur
Weak schur number.

Structs

NoFeatures
A move encoder that doesn't take into account the board state.

Traits

Base
Common interface for single and multiplayer games

Features
Games that can be represented as multi-dimensional arrays.

Game
Game with one or multiple players.

GameBuilder
Game builders.

GameView
Games with an user interface.

MoveCode
Move encoders

Playable
Mutable game by playing moves.

Playout
Game playouts

SingleWinner
Non-cooperative games.

Singleplayer
Single-player games.

SingleplayerGameBuilder
Builders for single-player games.

Functions

simulate
Simulate a match by executing the two policies on a given game.

Trait aliases

MoveTrait
Action that may be applied to a game state.