Module ggpf::game::breakthrough

Breakthrough game implementation

The rules are simple, a player wins when one of its pawns reaches the other side of the board.

Modules

ui
Breakthrough interactive interface.

Structs

Breakthrough
Breakthrough game state instance

BreakthroughBuilder
Game builder for Breakthough.

Move
Move

Enums

Cell
Game cell

Color
Players

MoveDirection
Move direction