Trait ggpf::game::Base

pub trait Base: Sized + Debug + Send + Sync {
    type Move: MoveTrait;
    fn possible_moves(&self) -> Vec<Self::Move>;

    fn is_finished(&self) -> bool { ... }
}

Common interface for single and multiplayer games

Associated Types

type Move: MoveTrait

The type for a Move.

Loading content...
Required methods

fn possible_moves(&self) -> Vec<Self::Move>

Given the game state and turn, list possible actions to the current player. If the game has ended, no action should be available.

Loading content...
Provided methods

fn is_finished(&self) -> bool

Returns if the game has ended or not.

Loading content...