Trait ggpf::game::Playout

pub trait Playout: Game + Clone + Send {
#[must_use]    fn playout_history<'life0, 'async_trait>(
        &'life0 self, 
        pov: Self::Player
    ) -> Pin<Box<dyn Future<Output = (Self, Vec<(Self, Self::Move)>, f32)> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

#[must_use]    fn playout_board<'life0, 'async_trait>(
        &'life0 self, 
        pov: Self::Player
    ) -> Pin<Box<dyn Future<Output = (Self, f32)> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Game playouts

Provided methods

#[must_use]fn playout_history<'life0, 'async_trait>(
&'life0 self,
pov: Self::Player
) -> Pin<Box<dyn Future<Output = (Self, Vec<(Self, Self::Move)>, f32)> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Simulate a game execution using random moves until reaching a final state. It stores moves and state history, along with the total reward and the final state.

#[must_use]fn playout_board<'life0, 'async_trait>(
&'life0 self,
pov: Self::Player
) -> Pin<Box<dyn Future<Output = (Self, f32)> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Simulates a game execution using random moves until reaching a final state. It returns the total reward with the final state.

Loading content...